refactors the code
